코드정리, 이미지추가, fade 디버그 테스트

+++ DX11_Wizard_of_Legend/ENGINE_SOURCE/jsInput.cpp
@@ -77,8 +77,8 @@ namespace js
 			POINT mousePos = {};
 			GetCursorPos(&mousePos);
 			ScreenToClient(application.GetHwnd(), &mousePos);
-			mMousPosition.x = mousePos.x;
-			mMousPosition.y = mousePos.y;
+			mMousPosition.x = (float)mousePos.x;
+			mMousPosition.y = (float)mousePos.y;
 		}
 		else
 		{
